requirements.txt
rpi_ws281x>=5.0.0
paho-mqtt==1.6.1
numpy==1.24.3 